% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_attachment}
\alias{add_attachment}
\title{add attachment to email}
\usage{
add_attachment(mail_base, att_content = NULL, att_type = "html",
  att_filename = NULL, tobase64 = TRUE)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{att_content}{string, path and filename}

\item{att_type}{string, type of content, default 'html'}

\item{att_filename}{string, filename}

\item{tobase64}{logical. By default attachment files converted to base64 inside function. Set to FALSE otherwise}
}
\description{
add attachment to email to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_attachment(my_mail, 'picture.png')
}
\keyword{email}
\keyword{sendgrid}
