% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cc.R
\name{add_cc}
\alias{add_cc}
\title{add "cc" email adresses}
\usage{
add_cc(mail_base, em_cc = NULL)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_cc}{list of characters}
}
\description{
add "cc" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_to(my_mail, 'mail1@foo.com')
}
\keyword{email}
\keyword{sendgrid}
