% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_content.R
\name{add_content}
\alias{add_content}
\title{add "content" email adresses}
\usage{
add_content(mail_base, em_content = NULL, content_type = "text/plain")
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_content}{list of characters}
}
\description{
add "content" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_content(my_mail, 'Hello world!')
}
\keyword{email}
\keyword{sendgrid}
