% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_from.R
\name{add_from}
\alias{add_from}
\title{add "from" email adresses}
\usage{
add_from(mail_base, em_from = NULL)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_from}{list of characters}
}
\description{
add "from" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_from(my_mail, 'mail1@foo.com')
}
\keyword{email}
\keyword{sendgrid}
