% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subject.R
\name{add_subject}
\alias{add_subject}
\title{add "subject" email adresses}
\usage{
add_subject(mail_base, em_subject = NULL)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_subject}{string}
}
\description{
add "subject" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_subject(my_mail, 'mail1@foo.com')
}
\keyword{email}
\keyword{sendgrid}
