% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to.R
\name{add_to}
\alias{add_to}
\title{add "to" email adresses}
\usage{
add_to(mail_base, to_mail)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_to}{list of characters}
}
\description{
cadd "to" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_to(my_mail, 'mail1@foo.com')
}
\keyword{email}
\keyword{sendgrid}
