% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_mail.R
\name{e_mail}
\alias{e_mail}
\title{Make json list of emails}
\usage{
e_mail(...)
}
\arguments{
\item{...}{list of characters}
}
\description{
Make json list of emails from caracter vector
 strings test for email format to be added
}
\examples{
msg_txt('hello','world!','','goodbye!')
}
\keyword{or}
\keyword{regex}
\keyword{string}
