% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_base.R
\name{email_base}
\alias{email_base}
\title{create base of email}
\usage{
email_base(em_to = "", em_cc = "", em_from = "", em_subject = "",
  content_type = "", content_value = "", att_content = "",
  att_type = "", att_filename = "")
}
\arguments{
\item{em_to}{list of characters}

\item{em_cc}{prefix, to be added on beginning of string}

\item{em_from}{suffix, to be added on end of the strin}

\item{em_subject}{separator, in normal paste(), here used as 'collapse' argument}
}
\description{
create base of email
Email components may be added on creation or later
}
\examples{
email_base()
}
\keyword{email}
\keyword{sendgrid}
