% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_txt.R
\name{msg_txt}
\alias{msg_txt}
\title{compose message body from character lines}
\usage{
msg_txt(...)
}
\arguments{
\item{...}{list of characters}
}
\description{
Takes chracter lines as input and paste it together with 2 linebreacks
 Use black string (i.e. '') to add extra line
}
\examples{
msg_txt('hello','world!','','goodbye!')
}
\keyword{email}
\keyword{string}
\keyword{text}
