% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendgrid_json_emails.R
\name{sendgrid_json_emails}
\alias{sendgrid_json_emails}
\title{Make json list of emails}
\usage{
sendgrid_json_emails(input)
}
\arguments{
\item{input}{character or list of characters}
}
\description{
Make json list of emails from caracter vector
 strings test for email format to be added
}
\details{
partly from https://github.com/rhochreiter/sendgridR/
}
\examples{
sendgrid_json_emails(c('bar@example.com', 'foo@example.com'))
# add name for email address with list
sendgrid_json_emails(list(list('bar@example.com', ' bar_bar'), 'foo@example.com'))
}
\keyword{email}
\keyword{sendgrid}
\keyword{string}
