% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sendgrid_message}
\alias{sendgrid_message}
\title{finalize message as request body}
\usage{
sendgrid_message(mail)
}
\arguments{
\item{mail}{list, mail created}
}
\description{
Takes chracter lines as input and paste it together with 2 linebreacks
 Use black string (i.e. '') to add extra line
}
\examples{
sendgrid_message(my_mail)
}
\keyword{email}
\keyword{string}
\keyword{text}
