% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_cc}
\alias{add_cc}
\title{add "cc" email adresses}
\usage{
add_cc(mail_base, cc_mail)
}
\arguments{
\item{mail_base}{base of email, created by email_base()}

\item{em_cc}{list of characters}
}
\description{
cadd "cc" email adresses to email base created earlier by email_base()
}
\examples{
my_mail <- email_base()
add_cc(my_mail, 'mail2@foo.com')
}
\keyword{email}
\keyword{sendgrid}
